+++ glean-core/ios/Glean/Testing/ErrorType.swift
@@ -6,15 +6,15 @@ import Foundation
 
 /// Enumeration of different metric lifetimes.
 public enum ErrorType: Int32 {
-    // For when the value to be recorded does not match the metric-specific restrictions
+    /// For when the value to be recorded does not match the metric-specific restrictions
     case invalidValue = 0
 
-    // For when the label of a labeled metric does not match the restrictions
+    /// For when the label of a labeled metric does not match the restrictions
     case invalidLabel = 1
 
-    // For when timings are recorded incorrectly
+    /// For when timings are recorded incorrectly
     case invalidState = 2
 
-    // For when the value to be recorded overflows the metric-specific upper range
+    /// For when the value to be recorded overflows the metric-specific upper range
     case invalidOverflow = 3
 }