+++ glean-core/ios/Glean/Utils/Utils.swift
@@ -22,6 +22,7 @@ extension UInt8 {
 ///
 /// This should only be used in tests.
 extension String: Error {
+    /// The string itself is the error description.
     public var errorDescription: String? { return self }
 }
 
@@ -202,11 +203,13 @@ func timestampNanos() -> UInt64 {
 }
 
 /// Gets a gecko-compatible locale string (e.g. "es-ES")
+// If the locale can't be determined on the system, the value is "und",
+// to indicate "undetermined".
 ///
 /// - returns: a locale string that supports custom injected locale/languages.
 func getLocaleTag() -> String {
     if NSLocale.current.languageCode == nil {
-        return "??"
+        return "und"
     } else {
         if NSLocale.current.regionCode == nil {
             return NSLocale.current.languageCode!