+++ glean-core/ios/GleanTests/Metrics/TimingDistributionMetricTests.swift
@@ -162,4 +162,39 @@ class TimingDistributionTypeTests: XCTestCase {
         let snapshot = try! metric.testGetValue()
         XCTAssertEqual(3, snapshot.count)
     }
+
+    func testMeasureFunctionThrows() {
+        let metric = TimingDistributionMetricType(
+            category: "telemetry",
+            name: "timing_distribution",
+            sendInPings: ["store1"],
+            lifetime: .application,
+            disabled: false,
+            timeUnit: .millisecond
+        )
+
+        XCTAssertFalse(metric.testHasValue())
+
+        // Create a test function that throws an exception.
+        func testFunc() throws {
+            throw "invalid"
+        }
+
+        // Measure a few times. Nothing should be recorded.
+        for _ in 1 ... 3 {
+            // Perform the measurement
+            do {
+                _ = try metric.measure {
+                    try testFunc()
+                }
+
+                // The function throws, so this is unreachable
+                XCTAssert(false)
+            } catch {
+                // intentionally left empty
+            }
+        }
+
+        XCTAssertFalse(metric.testHasValue())
+    }
 }