+++ glean-core/preview/src/lib.rs
@@ -86,7 +86,8 @@ fn with_glean<F, R>(f: F) -> R
 where
     F: Fn(&Glean) -> R,
 {
-    let lock = global_glean().lock().unwrap();
+    let glean = global_glean().expect("Global Glean object not initialized");
+    let lock = glean.lock().unwrap();
     f(&lock)
 }
 
@@ -94,7 +95,8 @@ fn with_glean_mut<F, R>(f: F) -> R
 where
     F: Fn(&mut Glean) -> R,
 {
-    let mut lock = global_glean().lock().unwrap();
+    let glean = global_glean().expect("Global Glean object not initialized");
+    let mut lock = glean.lock().unwrap();
     f(&mut lock)
 }
 
@@ -140,7 +142,6 @@ fn initialize_core_metrics(
     if let Some(app_channel) = channel {
         core_metrics.app_channel.set(glean, app_channel);
     }
-    core_metrics.os.set(glean, system::OS.to_string());
     core_metrics.os_version.set(glean, "unknown".to_string());
     core_metrics
         .architecture